ensured uniqueness and refactored
