+++ packages/pages/src/common/src/assets/getAssetsFilepath.ts
@@ -1,11 +1,13 @@
 import { pathToFileURL } from "url";
-import path from "node:path";
 import { UserConfig } from "vite";
 import { import_ } from "./import.js";
+import { sanitizeSubpath } from "./sanitizeSubpath.js";
 
 /**
- * Determines the assets directory to use by checking
- * vite.config.json's assetDir or default to "assets".
+ * Determines the assets directory to use by checking vite.config.json's
+ * assetDir or default to "assets". Empty values or paths that would escape
+ * the output directory fall back to the default; absolute paths are sanitized
+ * into a safe relative subpath for Rollup output.
  * @param defaultAssetsDir the default directory for assets
  * @param viteConfigPath the path to vite.config.js
  */
@@ -25,23 +27,9 @@ export const determineAssetsFilepath = async (
 
 /**
  * Ensures the assets directory is a safe, relative subpath for Rollup output.
- * Falls back when the value is empty, absolute, or would escape the output dir.
+ * Falls back when the value is empty or would escape the output dir; absolute
+ * paths are sanitized into safe relative subpaths.
  */
 const sanitizeAssetsDir = (assetsDir: string, fallback: string): string => {
-  const trimmed = assetsDir.trim();
-  if (trimmed.length === 0) {
-    return fallback;
-  }
-
-  const withoutDrive = trimmed.replace(/^[a-zA-Z]:/, "");
-  const withoutLeading = withoutDrive.replace(/^[/\\]+/, "");
-  const normalized = path.posix
-    .normalize(withoutLeading.replace(/\\/g, "/"))
-    .replace(/^(\.\/)+/, "");
-
-  if (normalized === "" || normalized === "." || normalized.startsWith("..")) {
-    return fallback;
-  }
-
-  return normalized;
+  return sanitizeSubpath(assetsDir, fallback);
 };
